requirements.txt
beautifulsoup4
jinja2
openai
pydantic
python-dotenv
requests
streamlit